+++ src/common/generateStream.ts
@@ -1,0 +1,30 @@
+export const generateStream = async (
+  headers: Record<string, string>,
+  body: string,
+  url: string
+): Promise<AsyncIterable<string>> => {
+  const response = await fetch(url, {
+    method: "POST",
+    headers,
+    body,
+  });
+  if (response.status !== 200) throw new Error(response.status.toString());
+  if (!response.body) throw new Error("Response body does not exist");
+  return getIterableStream(response.body);
+};
+
+async function* getIterableStream(
+  body: ReadableStream<Uint8Array>
+): AsyncIterable<string> {
+  const reader = body.getReader();
+  const decoder = new TextDecoder();
+
+  while (true) {
+    const { value, done } = await reader.read();
+    if (done) {
+      break;
+    }
+    const decodedChunk = decoder.decode(value, { stream: true });
+    yield decodedChunk;
+  }
+}
